Split Sentry webhook route into per-concern modules
The route file held five concerns at once: HTTP wiring, the injected-deps
contract, the two per-resource flows, and the generic parse/drop
mechanics. Split them so each file has one job and reads on its own.

- webhook-context.ts: the SentryWebhookContext dependency contract.
- webhook-delivery.ts: generic parseAndValidateOrDrop + recordAndDrop.
- issue-handler.ts / installation-handler.ts: the two resource flows.
- webhooks.ts: now just route wiring, the signature gate, and dispatch
  (195 -> 59 lines).

No behavior change.
